# Conflicts:
#	src/main/java/uet/oop/bomberman/components/entities/EntitiesManager.java
#	src/main/java/uet/oop/bomberman/components/entities/enemy/Balloom.java
#	src/main/java/uet/oop/bomberman/components/entities/enemy/Oneal.java

+++ src/main/java/uet/oop/bomberman/components/entities/Entity.java
@@ -53,7 +53,7 @@ public abstract class Entity {
         return height;
     }
 
-    public void setLocation(int x, int y) {
+    public void setLocation(double x, double y) {
         this.x = x;
         this.y = y;
     }
